Add cloud function that creates a subscription doc
A signed in user that doesn't complete the subscription should have
subscription status incomplete. We need to keep the subscription doc
private to everyone so the only place to create it is in a cloud
function.

There is no v2 auth onCreate method so I used the v2 identity
beforeUserCreated function, which is a blocking function. This meant
having to upgrade the project to use GCIP the new identity platform.